Restore chromakey effect
